+++ assets/javascripts/bootstrap-global-this-define.js
@@ -1,6 +1,0 @@
-// Set a `globalThis` so that bootstrap components are defined on window.bootstrap instead of window.
-window['bootstrap'] = {
-  "@popperjs/core": window.Popper,
-  _originalGlobalThis: window['globalThis']
-};
-window['globalThis'] = window['bootstrap'];