+++ assets/javascripts/bootstrap-global-this-undefine.js
@@ -1,2 +1,0 @@
-window['globalThis'] = window['bootstrap']._originalGlobalThis;
-window['bootstrap']._originalGlobalThis = null;