+++ lib/bootstrap.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'bootstrap/version'
-require 'popper_js'
 
 module Bootstrap
   class << self