+++ test/rails_test.rb
@@ -2,8 +2,11 @@ require 'test_helper_rails'
 
 class RailsTest < ActionDispatch::IntegrationTest
   include ::DummyRailsIntegration
+  include ::SassEngineSupport
 
   def test_visit_root
+    skip_unless_sass_can_compile_bootstrap!
+
     visit root_path
     # ^ will raise on JS errors
 
@@ -13,6 +16,8 @@ class RailsTest < ActionDispatch::IntegrationTest
   end
 
   def test_precompile
+    skip_unless_sass_can_compile_bootstrap!
+
     Dummy::Application.load_tasks
     Rake::Task['assets:precompile'].invoke
   end